Crate ejercicio_rust

Functions

lcs

let seq1 and se2 be sequences let m be the length of seq1, and let n be the length of seq2 Returns a grid with the lengths of subsequences.

main

print_diff

‘grid’ is the computed grid by lcs(seq1: &String, seq2: &String) seq1 and seq2 are the sequences i y j specify the location within ‘grid’ that we want to search for when diff is read. Initially we call the fn with i = len(m) j = len(n).

read_file_lines

Returns a Vec containing the lines of the file.

read_lines

The output is wrapped in a Result to allow matching on errors Returns an Iterator to the Reader of the lines of the file.